Moved CSS named-color table to gosub_shared so inline styles can resolve named colors
