+++ crates/gosub_css3/Cargo.toml
@@ -9,7 +9,6 @@ description = "CSS3 parser for Gosub"
 [dependencies]
 gosub_shared = { version = "0.1.1", registry = "gosub", path = "../gosub_shared", features = [] }
 gosub_interface = { version = "0.1.1", registry = "gosub", path = "../gosub_interface", features = [] }
-lazy_static = { workspace = true }
 log = { workspace = true }
 anyhow = { workspace = true, features = [] }
 colors-transform = "0.2.11"
